Convert LegendSpec.js -> Legend.spec.tsx
